requirements.txt
fastapi[standard]
httpx
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
